requirements.txt
tinydb==3.15.2
python-dotenv==0.12.0
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
python-crontab==2.4.1
python-dateutil==2.8.1
python-http-client==3.2.7
requests==2.23.0
sendgrid==6.2.1
six==1.14.0
urllib3==1.25.8
